Refactor resource pools and calculations to use uints and FP
* Change all old data types in configs that used double/float, to FP
* Change resource pool capacity from ulong to uint, more fit for purpose
* Adjusted all logic methods accordingly to new types
* Added QuantumHelper new method 'PowFp', and removed 'Pow' method from QuantumStatCalculator

+++ Assets/Src/FirstLight/Game/Data/DataTypes/ResourcePoolData.cs
@@ -10,10 +10,10 @@ namespace FirstLight.Game.Data.DataTypes
 	public struct ResourcePoolData : IEquatable<ResourcePoolData>
 	{
 		public GameId Id;
-		public ulong CurrentResourceAmountInPool;
+		public uint CurrentResourceAmountInPool;
 		public DateTime LastPoolRestockTime;
 		
-		public ResourcePoolData(GameId id, ulong currentResourceAmountInPool, DateTime lastPoolRestockTime)
+		public ResourcePoolData(GameId id, uint currentResourceAmountInPool, DateTime lastPoolRestockTime)
 		{
 			Id = id;
 			CurrentResourceAmountInPool = currentResourceAmountInPool;
